requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-engineio==4.12.0
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
wsproto==1.2.0
